Home screen implementation with Square Workflow.
