+++ ui-home/build.gradle.kts
@@ -28,9 +28,6 @@ dependencies {
     // Coroutines
     implementation(libraries.kotlinx.coroutines.core)
 
-    // FlowBinding
-    implementation(libraries.flowbinding.swipeRefreshLayout)
-
     // AndroidX
     implementation(libraries.androidx.swipeRefreshLayout)
     implementation(libraries.androidx.recyclerView)
@@ -47,7 +44,7 @@ dependencies {
     // Unit tests
     testImplementation(libraries.junit)
     testImplementation(libraries.truth)
-    testImplementation(libraries.androidx.archCore.testing)
+    testImplementation(libraries.workflow.testing)
     testImplementation(project(":domain-testing"))
     testImplementation(project(":coroutines-test-ext"))
