+++ ui-reading-list/build.gradle.kts
@@ -42,7 +42,6 @@ dependencies {
     // Unit tests
     testImplementation(libraries.junit)
     testImplementation(libraries.truth)
-    testImplementation(libraries.androidx.archCore.testing)
     testImplementation(project(":coroutines-test-ext"))
 
     // Android tests
