+++ ui-settings/build.gradle.kts
@@ -41,7 +41,6 @@ dependencies {
     // Unit tests
     testImplementation(libraries.junit)
     testImplementation(libraries.truth)
-    testImplementation(libraries.androidx.archCore.testing)
     testImplementation(project(":coroutines-test-ext"))
 
     // Android tests
