+++ ui-story-details/build.gradle.kts
@@ -49,7 +49,6 @@ dependencies {
     // Unit tests
     testImplementation(libraries.junit)
     testImplementation(libraries.truth)
-    testImplementation(libraries.androidx.archCore.testing)
     testImplementation(project(":coroutines-test-ext"))
 
     // Android tests
